Create Run a Show command on a Cisco switch
Once in Ansible you will need to use nano or some other text editor to edit the "hosts" file. I made a group called [ios] and added the IP addresses of the devices I wanted to query. Line 3 of the script calls on all devices listed under the [ios] group.